+++ src/index.html
@@ -18,7 +18,7 @@
 <!--            <option value="gravity" title="Gravity">Gravity</option>-->
             <option value="gravityPro" title="Gravity Pro">Gravity Pro</option>
             <option value="pulsar" title="Pulsar">Pulsar</option>
-            <option value="cluster3" title="ClusterPro">Cluster Pro</option>
+            <option value="cluster3" title="ClusterPro">Cluster Pro Plus</option>
             <option value="transient" title="Transient">Transient</option>
           </select>
         </form>
